Enabled schema validation and now check for empty string parameters
I also changed the licencePlateCountryID input parameter for the 
contractCreateTest operation from "int" to "Integer". This was necessary
because this is an optional parameter and if the parameter is not 
supplied, we need to be able to have "null" supplied as a value (which
is not possible for "int"s).